+++ lib/src/controller/survey_controller.dart
@@ -5,14 +5,44 @@ import 'package:survey_kit/src/presenter/survey_presenter.dart';
 import 'package:survey_kit/src/result/question_result.dart';
 
 class SurveyController {
+  /// Defines what should happen if the next step is called
+  /// Default behavior is:
+  /// ```dart
+  /// BlocProvider.of<SurveyPresenter>(context).add(
+  ///    NextStep(
+  ///      resultFunction.call(),
+  ///    ),
+  /// );
+  /// ```
   final Function(
     BuildContext context,
     QuestionResult Function() resultFunction,
   )? onNextStep;
+
+  /// Defines what should happen if the previous step is called
+  /// Default behavior is:
+  /// ```dart
+  /// BlocProvider.of<SurveyPresenter>(context).add(
+  ///    StepBack(
+  ///      resultFunction.call(),
+  ///    ),
+  /// );
+  /// ```
   final Function(
     BuildContext context,
     QuestionResult Function() resultFunction,
   )? onStepBack;
+
+  /// Defines what should happen if the survey should be closed
+  /// Default behavior is:
+  /// ```dart
+  /// BlocProvider.of<SurveyPresenter>(context).add(
+  ///    CloseSurvey(
+  ///      resultFunction.call(),
+  ///    ),
+  /// );
+  /// Navigator.pop(context);
+  /// ```
   final Function(
     BuildContext context,
     QuestionResult Function() resultFunction,