+++ lib/src/answer_format/multiple_choice_answer_format.dart
@@ -1,12 +1,21 @@
+import 'package:json_annotation/json_annotation.dart';
 import 'package:survey_kit/src/answer_format/answer_format.dart';
 import 'package:survey_kit/src/answer_format/text_choice.dart';
 
+part 'multiple_choice_answer_format.g.dart';
+
+@JsonSerializable()
 class MultipleChoiceAnswerFormat implements AnswerFormat {
   final List<TextChoice> textChoices;
+  @JsonKey(defaultValue: const [])
   final List<TextChoice> defaultSelection;
 
-  MultipleChoiceAnswerFormat({
+  const MultipleChoiceAnswerFormat({
     required this.textChoices,
     this.defaultSelection = const [],
-  });
+  }) : super();
+
+  factory MultipleChoiceAnswerFormat.fromJson(Map<String, dynamic> json) =>
+      _$MultipleChoiceAnswerFormatFromJson(json);
+  Map<String, dynamic> toJson() => _$MultipleChoiceAnswerFormatToJson(this);
 }