+++ lib/src/answer_format/time_answer_formart.dart
@@ -1,10 +1,41 @@
 import 'package:flutter/material.dart';
+import 'package:json_annotation/json_annotation.dart';
 import 'package:survey_kit/src/answer_format/answer_format.dart';
 
+part 'time_answer_formart.g.dart';
+
+@JsonSerializable()
 class TimeAnswerFormat implements AnswerFormat {
+  @_TimeOfDayJsonConverter()
   final TimeOfDay? defaultValue;
 
-  TimeAnswerFormat({
+  const TimeAnswerFormat({
     this.defaultValue,
-  });
+  }) : super();
+
+  factory TimeAnswerFormat.fromJson(Map<String, dynamic> json) =>
+      _$TimeAnswerFormatFromJson(json);
+  Map<String, dynamic> toJson() => _$TimeAnswerFormatToJson(this);
+}
+
+class _TimeOfDayJsonConverter
+    implements JsonConverter<TimeOfDay?, Map<String, dynamic>> {
+  const _TimeOfDayJsonConverter();
+
+  @override
+  TimeOfDay? fromJson(Map<String, dynamic> json) {
+    if (json['hour'] == null || json['minute'] == null) {
+      return null;
+    }
+    return TimeOfDay(
+      hour: json['hour'] as int,
+      minute: json['minute'] as int,
+    );
+  }
+
+  @override
+  Map<String, dynamic> toJson(TimeOfDay? timeOfDay) => <String, dynamic>{
+        'hour': timeOfDay?.hour,
+        'minute': timeOfDay?.minute,
+      };
 }