+++ lib/src/steps/step_not_defined_exception.dart
@@ -1,0 +1,3 @@
+class StepNotDefinedException implements Exception {
+  const StepNotDefinedException() : super();
+}