+++ lib/src/task/navigable_task.dart
@@ -5,12 +5,14 @@ import 'package:survey_kit/src/task/task.dart';
 import 'package:survey_kit/src/task/identifier/task_identifier.dart';
 
 class NavigableTask extends Task {
-  Map<StepIdentifier, NavigationRule> navigationRules = {};
+  final Map<StepIdentifier, NavigationRule> navigationRules;
 
   NavigableTask({
     TaskIdentifier? id,
     List<Step> steps = const [],
-  }) : super(id: id, steps: steps);
+    Map<StepIdentifier, NavigationRule>? navigationRules,
+  })  : this.navigationRules = navigationRules ?? {},
+        super(id: id, steps: steps);
 
   void addNavigationRule(
       {required StepIdentifier forTriggerStepIdentifier,
@@ -20,5 +22,24 @@ class NavigableTask extends Task {
 
   NavigationRule? getRuleByStepIdentifier(StepIdentifier? stepIdentifier) {
     return navigationRules[stepIdentifier];
+  }
+
+  factory NavigableTask.fromJson(Map<String, dynamic> json) {
+    Map<StepIdentifier, NavigationRule> navigationRules = {};
+    final rules = json['rules'] as List;
+    rules.forEach((rule) {
+      print(rule);
+      navigationRules.putIfAbsent(
+          StepIdentifier.fromJson(rule['triggerStepIdentifier']),
+          () => NavigationRule.fromJson(rule as Map<String, dynamic>));
+    });
+
+    return NavigableTask(
+      id: TaskIdentifier.fromJson(json),
+      steps: (json['steps'] as List)
+          .map((step) => Step.fromJson(step as Map<String, dynamic>))
+          .toList(),
+      navigationRules: navigationRules,
+    );
   }
 }