+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - BREAKING: Migrated to null-safety
 - Upgrade Dart SDK constraints to >=2.12.0-0 <3.0.0
+- Expose SurveyController to override navigation (StepBack, NextStep and CloseSurvey)
+- Flutter SurveyKit now also used with Web, MacOS, Linux, Windows (Not optimized)
+- Updated to flutter_platform_widgets
 
 # 0.0.2 - 0.0.8