+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - BREAKING: Migrated to null-safety
 - BREAKING: Upgrade Dart SDK constraints to >=2.12.0-0 <3.0.0
-- BREAKING: Expose SurveyController to override navigation (StepBack, NextStep and CloseSurvey)
+- BREAKING: Expose SurveyController to add the possiblity to override the navigation (StepBack, NextStep and CloseSurvey)
 - Flutter SurveyKit now also used with Web, MacOS, Linux, Windows (Not optimized)
 - Updated to flutter_platform_widgets